requirements.txt
asgiref==3.11.1
Django==5.2.13
djangorestframework==3.17.1
sqlparse==0.5.5
typing_extensions==4.15.0
